+++ R/obtain_exclusive.R
@@ -2,18 +2,15 @@
 #'
 #' @description Get exclusive proteins/sites (only used when filter.protein.type = "fraction")
 #'
-#' @param data Input or modified proteome file from MaxQuant (.txt)
-#' @param Fraction Can be either "Proteome" or "Enriched". Indicates the type of input data used
-#' @param sampleTable .xlsx file containing information about the samples. Three columns are mandatory (label, condition and replicate)
-#' @param contrasts Mentions the conditions to be compared. Ex. MUTANT_vs_WILDTYPE or (MUTANT-A_vs_WILDTYPE-A)_vs_(MUTANT-B_vs_WILDTYPE-B). This how the contrasts should be provided and the values should be the same as the one given in condition column of sampleTable. (MUTANT-A_vs_WILDTYPE-A)_vs_(MUTANT-B_vs_WILDTYPE-B): This type can be used for complex data when comparing the interaction between two conditions such genotype and time
-#'
+#' @param data output file (or output) from editData function
+#
 #' @return List of exclsive proteins/sites from all comparisons (except difference of difference comparisons)
 #'
 #' @export
 #'
 ##################### Obtain exclusive proteins and sites ########################
 
-obtain_exclusive <- function(data, Fraction, sampleTable, contrasts){
+obtain_exclusive <- function(data){
 
   data.na <- data